backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
asyncpg
psycopg[binary]
redis
google-genai
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
httpx
lancedb
